requirements.txt
discord.py
pyyaml
pyyaml-include